basic orbit simulation system
this is the basic foundation of the project, having bodies with different masses attracting and influencing each others velocity. kinda hard to figure out the starting values for a stable configuration, but i did do it with a single planet and a sun. this stuff uses newtons law of gravitation (F = G(m1m2)/R2). my next step will be to add a visualisation of the orbits.