orbits
it runs kind of slow with the orbits as its drawing from 10k points every frame for every planet(not sun), but i guess it works. it runs more smoothly with like 5k points instead, but the orbit is shorter. i guess this can be used to figure out the initial values for position and velocity of the bodies, and then running the program after youve found a good uniform orbit. the way this works is that in setup, for a specified number of steps (10k at default), it runs the simulation with all the bodies, and puts their position after each step into their own "orbit" array. after the 10k steps, their position and velocities are set back to whatever they were at the beginning. and then every frame in the display function of bodies, it also draws the orbit by drawing from all the 10k positions in the orbit array for the corresponding body.