actual solar system
theres 5 planets, and all 5 have almost a circular orbit around the sun.

although i kind of cheated here by setting the masses of all the planets to be way lower, so theyre not affecting each others orbits as much as they are getting affected by the sun themselves.

but it works